[ADD] adding generic base serializers that will be inherited by interface-specific applications such as API and Website; adding the BaseSerializer, base serializer fields such as source and url fields; adding the relationship base serializers
