--- src/interfaces/serializers/__init__.py
@@ -1,34 +1,0 @@
-from .datasets import (TRNListSerializer,
-                       TRNDetailSerializer,
-                       OrganismBindingSitesListSerializer,
-                       OrganismBindingSitesDetailSerializer,
-                       RegulatorBindingSitesListSerializer,
-                       RegulatorBindingSitesDetailSerializer)
-from .effector import (EffectorDetailSerializer,
-                       EffectorField,
-                       EffectorListSerializer)
-from .evidence import (EvidenceDetailSerializer,
-                       EvidenceListSerializer)
-from .gene import (GeneDetailSerializer,
-                   GeneField,
-                   GeneListSerializer)
-from .operon import (OperonDetailSerializer,
-                     OperonListSerializer)
-from .organism import (OrganismDetailSerializer,
-                       OrganismField,
-                       OrganismListSerializer,
-                       OrganismsSerializer)
-from .pathway import (PathwayDetailSerializer,
-                      PathwayListSerializer)
-from .publication import (PublicationDetailSerializer,
-                          PublicationListSerializer)
-from .regulator import (RegulatorDetailSerializer,
-                        RegulatorField,
-                        RegulatorListSerializer)
-from .regulatory_family import (RegulatoryFamilyDetailSerializer,
-                                RegulatoryFamilyListSerializer)
-from .regulatory_interaction import (RegulatoryInteractionDetailSerializer,
-                                     RegulatoryInteractionListSerializer)
-from .tfbs import (TFBSDetailSerializer,
-                   TFBSField,
-                   TFBSListSerializer)
